+++ src/main/kotlin/com/my/sparta/concert/aggregate/user/adapter/inbound/web/interfaces/response/UserWalletInfoResponse.kt
@@ -1,9 +1,10 @@
 package com.my.sparta.concert.aggregate.user.adapter.inbound.web.interfaces.response
 
 import lombok.Data
+import java.math.BigDecimal
 
 @Data
 class UserWalletInfoResponse(
     val userId: String,
-    val money: Double,
+    val money: BigDecimal,
 )
